Move set default coins for vault to keygen and import vault
