+++ VultisigApp/VultisigApp/View Models/VaultDetailViewModel.swift
@@ -12,13 +12,8 @@ import SwiftUI
 class VaultDetailViewModel: ObservableObject {
     @Published var coinsGroupedByChains = [GroupedChain]()
     @Published var selectedGroup: GroupedChain? = nil
-    private let semaphore = DispatchSemaphore(value: 1)
-    
-    let defaultChains = [Chain.bitcoin, Chain.ethereum, Chain.thorChain, Chain.solana]
     let balanceService = BalanceService.shared
-    
     private var updateBalanceTask: Task<Void, Never>?
-    
     func updateBalance() {
         updateBalanceTask?.cancel()
         updateBalanceTask = Task {
@@ -32,50 +27,7 @@ class VaultDetailViewModel: ObservableObject {
             coinsGroupedByChains[index].setOrder(index)
         }
     }
-    func setDefaultCoinsOnce(for vault: Vault) {
-        semaphore.wait()
-        Task{
-            await setDefaultCoins(for: vault)
-            semaphore.signal()
-        }
-    }
-    func setDefaultCoins(for vault: Vault) async {
-        // add bitcoin when the vault doesn't have any coins in it
-        if vault.coins.count == 0 {
-            print("set default coins for vault:\(vault.name)")
-            for chain in defaultChains {
-                var result: Result<Coin,Error>
-                switch chain {
-                case .bscChain:
-                    result = EVMHelper(coinType: .smartChain).getCoin(hexPubKey: vault.pubKeyEdDSA, hexChainCode: vault.hexChainCode)
-                case .bitcoin:
-                    result = UTXOChainsHelper(coin: .bitcoin, vaultHexPublicKey: vault.pubKeyECDSA, vaultHexChainCode: vault.hexChainCode).getCoin()
-                case .ethereum:
-                    result = EVMHelper(coinType: .ethereum).getCoin(hexPubKey: vault.pubKeyECDSA, hexChainCode: vault.hexChainCode)
-                case .thorChain:
-                    result = THORChainHelper.getRUNECoin(hexPubKey: vault.pubKeyECDSA, hexChainCode: vault.hexChainCode)
-                case .solana:
-                    result = SolanaHelper.getSolana(hexPubKey: vault.pubKeyEdDSA, hexChainCode: vault.hexChainCode)
-                default:
-                    continue
-                }
-                
-                switch result {
-                case .success(let btc):
-                    do{
-                        try await Storage.shared.save(btc)
-                        vault.coins.append(btc)
-                    }catch{
-                        print("fail to save coin: \(error)")
-                    }
-                    
-                case .failure(let error):
-                    print("error: \(error)")
-                }
-                
-            }
-        }
-    }
+    
     
     func fetchCoins(for vault: Vault) {
         categorizeCoins(vault: vault)
